fix administrator DAO
